Updated CRUD methods for Division

+++ EmployeeDashboardDemo/Controllers/DivisionsController.cs
@@ -40,24 +40,23 @@ namespace EmployeeDashboardDemo.Controllers
             return View(division);
         }
 
-        // GET: Divisions/Create
-        public ActionResult Create()
+        // GET: Divisions/Add
+        public ActionResult Add()
         {
             return View();
         }
 
-        // POST: Divisions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Divisions/Add
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,DateCreated")] Division division)
+        public ActionResult Add(Division division)
         {
             if (ModelState.IsValid)
             {
                 db.Divisions.Add(division);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                TempData["Success"] = "Division added successfully.";
+                return RedirectToAction("Add");
             }
 
             return View(division);
@@ -67,30 +66,28 @@ namespace EmployeeDashboardDemo.Controllers
         public ActionResult Edit(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Division division = db.Divisions.Find(id);
             if (division == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(division);
         }
 
         // POST: Divisions/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,DateCreated")] Division division)
+        public ActionResult Edit([Bind(Include = "Id,Name")] Division division)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(division).State = EntityState.Modified;
                 db.SaveChanges();
+                TempData["Success"] = "Division updated successfully.";
                 return RedirectToAction("Index");
             }
+
             return View(division);
         }
 
@@ -98,15 +95,13 @@ namespace EmployeeDashboardDemo.Controllers
         public ActionResult Delete(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Division division = db.Divisions.Find(id);
             if (division == null)
-            {
                 return HttpNotFound();
-            }
-            return View(division);
+
+            return View(division); // confirmation page
         }
 
         // POST: Divisions/Delete/5
@@ -115,8 +110,13 @@ namespace EmployeeDashboardDemo.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Division division = db.Divisions.Find(id);
-            db.Divisions.Remove(division);
-            db.SaveChanges();
+            if (division != null)
+            {
+                db.Divisions.Remove(division);
+                db.SaveChanges();
+                TempData["Success"] = "Division deleted successfully.";
+            }
+
             return RedirectToAction("Index");
         }
